[#170] refactor: useNavigateWithAuth 훅을 별도 파일로 분리
- src/hooks/navigate.ts 파일 생성
- useNavigateWithAuth 훅과 NavigateCallback 타입을 common.ts에서 navigate.ts로 이동
- 관련 import 경로 업데이트 (FloatingActionMenu, useNavigateWithAuth 테스트)
- 관심사 분리를 통한 코드 구조 개선
