+++ src/hooks/navigate.ts
@@ -1,0 +1,20 @@
+import { useCallback } from "react";
+import { useRouter } from "next/navigation";
+
+import { loginCheck } from "@/utils/common";
+
+type NavigateCallback = () => void;
+
+export const useNavigateWithAuth = () => {
+  const router = useRouter();
+
+  return useCallback(
+    (path: string, onNavigate?: NavigateCallback) => {
+      if (loginCheck()) {
+        router.push(path);
+        onNavigate?.();
+      }
+    },
+    [router]
+  );
+};
